# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentResolver — X402 Settlement Canary & Payment Preflight",
    "version": "0.1.6",
    "description": "Machine-readable x402 settlement testing and transaction-path verification for autonomous buyers. For the lowest-friction first paid integration, GET /api/x402-ping settles $0.001 USDC on Base or Solana and returns a timestamped delivery canary. Before authorizing a target x402 purchase, POST /api/x402-payment-preflight checks endpoint reachability, PAYMENT-REQUIRED structure, payTo recipient, quoted price, network, asset, exact scheme and resource binding, then returns a fail-closed eligible/blocked decision and reusable evidence fingerprints. Free capability resolution remains available as a secondary surface. Provider-side distribution is zero-account: a catalog-discovered x402 provider may publish a same-origin AgentResolver well-known manifest and opt into a 2% verified success fee with a $0.001 minimum; buyers pay no extra AgentResolver fee.",
...
  },
  "servers": [
    {
      "url": "https://agentresolver.vercel.app"
    }
  ],
...
  "paths": {
...
    "/api/procure": {
      "post": {
        "operationId": "procureCapability",
        "tags": [
          "Agent Capability Procurement"
        ],
        "summary": "Procure a compatible machine capability under hard constraints",
        "description": "Free, read-only and non-custodial. Declare the capability needed plus budget/network/protocol/schema/side-effect/auth constraints. AgentResolver can procure across x402, L402, MPP, MCP and ordinary HTTPS capabilities, rejects incompatible candidates, surfaces unknown evidence explicitly, and returns a direct execution/payment handoff. It never receives a private key or authorizes caller spend. A caller may also seed at most two already-known provider origins; these are usable only after same-origin manifest and live x402 payment-identity verification.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "goal"
                ],
                "additionalProperties": false,
                "properties": {
                  "goal": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000
                  },
                  "limit": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 20,
                    "default": 5
                  },
                  "constraints": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "maxPriceUsd": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 1000
                      },
                      "preferredNetworks": {
                        "type": "array",
                        "maxItems": 8,
                        "items": {
                          "type": "string"
                        }
                      },
                      "protocol": {
                        "type": "string",
                        "enum": [
                          "x402",
                          "l402",
                          "mpp",
                          "mcp",
                          "any"
                        ]
                      },
                      "requireHttps": {
                        "type": "boolean",
                        "default": true
                      },
                      "availableInputSchema": {
                        "type": "object"
                      },
                      "requiredOutputSchema": {
                        "type": "object"
                      },
                      "sideEffect": {
                        "type": "string",
                        "enum": [
                          "read-only",
                          "state-changing",
                          "any"
                        ]
                      },
                      "auth": {
                        "type": "string",
                        "enum": [
                          "none",
                          "wallet",
                          "api-key",
                          "any"
                        ]
                      }
                    }
                  },
                  "task": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000,
                    "description": "Alias for goal."
                  },
                  "query": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000,
                    "description": "Alias for goal."
                  },
                  "q": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000,
                    "description": "Alias for goal."
                  },
                  "providerOrigins": {
                    "type": "array",
                    "maxItems": 2,
                    "uniqueItems": true,
                    "items": {
                      "type": "string",
                      "format": "uri",
                      "pattern": "^https://[^/]+/?$"
                    },
                    "description": "Optional provider HTTPS origins already known to the caller. AgentResolver reads only the fixed /.well-known/agentresolver-provider.json manifest and live-verifies matching same-origin x402 routes before they can enter this procurement call."
                  }
                }
              },
              "examples": {
                "x402Search": {
                  "value": {
                    "goal": "search the public web and return structured results",
                    "constraints": {
                      "maxPriceUsd": 0.05,
                      "protocol": "x402",
                      "preferredNetworks": [
                        "eip155:8453"
                      ],
                      "requireHttps": true
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked compatible candidates, explicit rejection/unknown reasons, and a non-custodial execution/payment handoff."
          },
          "400": {
            "description": "Invalid procurement request or constraint schema."
          }
        },
        "security": [],
        "x-agentresolver-supply-sources": [
          "AgentResolver owned capabilities",
          "caller-seeded verified provider origins",
          "domain-enrolled providers",
          "402 Index",
          "PayAI",
          "Circle x402 discovery",
          "MCP directories"
        ]
      },
      "get": {
        "operationId": "procureCapabilityGet",
        "tags": [
          "Agent Capability Procurement"
        ],
        "summary": "Procure a compatible capability with URL query parameters",
        "description": "Compatibility-first free procurement for agents and crawlers that cannot issue the JSON POST contract. Accepts goal, task, query, or q plus optional limit, maxPriceUsd, protocol, network/preferredNetwork, and requireHttps. Returns the same non-custodial procurement response as POST.",
        "parameters": [
          {
            "name": "goal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 1000
            }
          },
          {
            "name": "task",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 1000
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 1000
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 1000
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20,
              "default": 5
            }
          },
          {
            "name": "maxPriceUsd",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 1000
            }
          },
          {
            "name": "protocol",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "x402",
                "l402",
                "mpp",
                "mcp",
                "any"
              ]
            }
          },
          {
            "name": "network",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "preferredNetwork",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requireHttps",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "providerOrigin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "maxItems": 2,
              "items": {
                "type": "string",
                "format": "uri"
              }
            },
            "style": "form",
            "explode": true,
            "description": "Repeat up to twice to seed an already-known provider HTTPS origin. The fixed AgentResolver well-known manifest and listed route payment identity are live-verified before use."
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked compatible candidates and non-custodial execution/payment handoff."
          },
          "400": {
            "description": "No usable goal/query supplied or invalid constraint."
          }
        },
        "security": []
      }
    },
...
  }
}